fix: Rename MSGraphDbUtils to MSGraphDb
